Changes automysqlbackup from a singleton class to a defined type.
As it stands you can only have a single instance of automysqlbackup on
a single machine. This modifies it such that you can use the defined type
in your own modules, or pass a parameter called 'config' (and config_defaults)
to the automysqlclass to use the create_resources function to create as many
instances as necessary.

This allows it to be used entirely from hiera.